// src/app/shared/interface/nave-espacial.interface.ts
import { Personagem } from '../../shared/interface/personagem.interface';

export interface NaveEspacial {
    id: number;
    name: string;
    model: string;
    manufacturer: string;
    cost_in_credits: string;
    length: string;
    max_atmosphering_speed: string;
    crew: string;
    passengers: string;
    cargo_capacity: string;
    consumables: string;
    hyperdrive_rating: string;
    MGLT: string;
    starship_class: string;
    pilots: Personagem[];
    films: string[];
    created: Date;
    edited: Date;
    url: string;
}